// src/app/models/userDto.ts
export interface UserDto {
    
    /**
     * id de la personne
    */
    id?: number;
    /**
     * prenom de la personne
     */
    prenom?: string;
    /**
     * nom de la personne
     */
    nom?: string;
    /**
     * age de la personne
     */
    email?: string;
    /**
     * qualification de la personne
     */
    password?: string;
}